CFY-6006 Move old tests to the new test_blueprint file
